requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
alibabacloud_dysmsapi20170525
alibabacloud_tea_openapi 
alibabacloud_darabonba_env
alibabacloud_dysmsapi20170525 
alibabacloud_tea_util
alibabacloud_tea_console
alibabacloud_darabonba_string
alibabacloud_darabonba_time